Refactor OHOS validation tooling
- Move OHOS build validation, HAP inspection, and device smoke logic into focused modules.
- Move the CLI options and generated parser into the device-smoke module.
- Split OHOS tests by responsibility.

Mechanical method: apply_patch exact block extraction; regenerate with `cd tools/frb_internal && dart run build_runner build --delete-conflicting-outputs`.
