+++ tools/frb_internal/test/makefile_dart_test.dart
@@ -57,348 +57,6 @@ void main() {
     );
   });
 
-  test('OHOS Rust library name follows Cargo crate naming rules', () {
-    expect(
-      ohosRustLibraryNameForTesting('''
-[package]
-name = "rust-lib-example"
-version = "0.1.0"
-'''),
-      'librust_lib_example.so',
-    );
-    expect(
-      ohosRustLibraryNameForTesting('''
-[package]
-name = "rust-lib-example"
-version = "0.1.0"
-
-[lib]
-name = "custom_bridge"
-crate-type = ["cdylib"]
-'''),
-      'libcustom_bridge.so',
-    );
-  });
-
-  test('OHOS HAP validation requires the arm64 Rust dynamic library', () {
-    expect(
-      ohosHapContainsRustLibraryForTesting([
-        'libs/arm64-v8a/libflutter.so',
-        'libs/arm64-v8a/librust_lib_example.so',
-      ], expectedLibrary: 'librust_lib_example.so'),
-      isTrue,
-    );
-    expect(
-      ohosHapContainsRustLibraryForTesting([
-        'libs/x86_64/librust_lib_example.so',
-      ], expectedLibrary: 'librust_lib_example.so'),
-      isFalse,
-    );
-    expect(
-      ohosHapContainsRustLibraryForTesting([
-        'assets/arm64-v8a/librust_lib_example.so',
-      ], expectedLibrary: 'librust_lib_example.so'),
-      isFalse,
-    );
-    expect(
-      ohosHapContainsRustLibraryForTesting([
-        'libs/arm64-v8a/libflutter.so',
-      ], expectedLibrary: 'librust_lib_example.so'),
-      isFalse,
-    );
-  });
-
-  test('OHOS HAP validation resolves app and plugin Rust manifests', () {
-    expect(
-      ohosRustCargoTomlPathForTesting(
-        packageDir: '/workspace/app',
-        fileExists: (candidate) =>
-            candidate == '/workspace/app/rust/Cargo.toml',
-      ),
-      '/workspace/app/rust/Cargo.toml',
-    );
-    expect(
-      ohosRustCargoTomlPathForTesting(
-        packageDir: '/workspace/plugin/example',
-        fileExists: (candidate) =>
-            candidate == '/workspace/plugin/rust/Cargo.toml',
-      ),
-      '/workspace/plugin/rust/Cargo.toml',
-    );
-  });
-
-  test('OHOS HAP backup restores the previous output after failure', () {
-    final temporaryDirectory = Directory.systemTemp.createTempSync(
-      'frb_ohos_hap_backup_test_',
-    );
-    try {
-      final output = Directory('${temporaryDirectory.path}/hap')..createSync();
-      File('${output.path}/previous.hap').writeAsStringSync('previous');
-
-      final backup = stashOhosHapOutputForTesting(output);
-      expect(output.existsSync(), isFalse);
-      expect(backup.backup!.existsSync(), isTrue);
-
-      output.createSync();
-      File('${output.path}/failed.hap').writeAsStringSync('failed');
-      restoreOhosHapOutputForTesting(backup);
-
-      expect(
-        File('${output.path}/previous.hap').readAsStringSync(),
-        'previous',
-      );
-      expect(File('${output.path}/failed.hap').existsSync(), isFalse);
-      expect(backup.backup!.existsSync(), isFalse);
-    } finally {
-      temporaryDirectory.deleteSync(recursive: true);
-    }
-  });
-
-  test('OHOS HAP backup removes failed first-build output', () {
-    final temporaryDirectory = Directory.systemTemp.createTempSync(
-      'frb_ohos_hap_first_build_test_',
-    );
-    try {
-      final output = Directory('${temporaryDirectory.path}/hap');
-      final backup = stashOhosHapOutputForTesting(output);
-      expect(backup.backup, isNull);
-
-      output.createSync();
-      File('${output.path}/failed.hap').writeAsStringSync('failed');
-      restoreOhosHapOutputForTesting(backup);
-
-      expect(output.existsSync(), isFalse);
-    } finally {
-      temporaryDirectory.deleteSync(recursive: true);
-    }
-  });
-
-  test('OHOS HAP inspection falls back from JDK jar to unzip', () {
-    final candidates = ohosArchiveListerCandidatesForTesting();
-    expect(candidates.map((candidate) => candidate.executable), [
-      'jar',
-      'unzip',
-    ]);
-    expect(candidates[0].arguments, ['--version']);
-    expect(candidates[1].arguments, ['-v']);
-  });
-
-  test('OHOS SDK preflight accepts a complete native SDK directory', () {
-    const sdkHome = '/opt/ohos/18/native';
-    final existingPaths = {
-      sdkHome,
-      '$sdkHome/llvm/bin/clang',
-      '$sdkHome/llvm/bin/llvm-ar',
-      '$sdkHome/sysroot',
-    };
-
-    expect(
-      validateOhosSdkHomeForTesting(
-        sdkHome: sdkHome,
-        isWindows: false,
-        pathExists: existingPaths.contains,
-      ),
-      isEmpty,
-    );
-  });
-
-  test('OHOS SDK preflight reports missing native toolchain components', () {
-    const sdkHome = '/opt/ohos/18/native';
-
-    expect(
-      validateOhosSdkHomeForTesting(
-        sdkHome: sdkHome,
-        isWindows: false,
-        pathExists: (candidate) => candidate == sdkHome,
-      ),
-      containsAll([
-        contains('llvm/bin/clang'),
-        contains('llvm/bin/llvm-ar'),
-        contains('sysroot'),
-      ]),
-    );
-  });
-
-  test('OHOS SDK preflight rejects missing and unsafe SDK paths', () {
-    expect(
-      validateOhosSdkHomeForTesting(
-        sdkHome: null,
-        isWindows: false,
-        pathExists: (_) => false,
-      ).single,
-      contains('is not set'),
-    );
-
-    final errors = validateOhosSdkHomeForTesting(
-      sdkHome: '/opt/鸿蒙 sdk/native',
-      isWindows: false,
-      pathExists: (_) => true,
-    );
-    expect(errors, contains(contains('contains whitespace')));
-    expect(errors, contains(contains('contains non-ASCII')));
-  });
-
-  test('OHOS Flutter preflight only accepts ohos in platforms help', () {
-    expect(
-      ohosFlutterCreateHelpSupportsPlatformForTesting('''
---platforms          The platforms supported by this project.
-                     [android, ios, ohos]
---project-name       The project name.
-'''),
-      isTrue,
-    );
-    expect(
-      ohosFlutterCreateHelpSupportsPlatformForTesting('''
---platforms          The platforms supported by this project.
-                     [android, ios]
---description        Mentions OHOS elsewhere.
-'''),
-      isFalse,
-    );
-  });
-
-  test('OHOS device smoke resolves one connected target', () {
-    expect(
-      resolveOhosDeviceIdForTesting(
-        '3QC0124C20001268\n',
-        requestedDeviceId: null,
-      ),
-      '3QC0124C20001268',
-    );
-    expect(
-      () => resolveOhosDeviceIdForTesting('[Empty]\n', requestedDeviceId: null),
-      throwsStateError,
-    );
-    expect(
-      () => resolveOhosDeviceIdForTesting(
-        'device-a\ndevice-b\n',
-        requestedDeviceId: null,
-      ),
-      throwsStateError,
-    );
-  });
-
-  test('OHOS device smoke validates an explicitly selected target', () {
-    expect(
-      resolveOhosDeviceIdForTesting(
-        'device-a\ndevice-b\n',
-        requestedDeviceId: 'device-b',
-      ),
-      'device-b',
-    );
-    expect(
-      () => resolveOhosDeviceIdForTesting(
-        'device-a\n',
-        requestedDeviceId: 'device-b',
-      ),
-      throwsStateError,
-    );
-  });
-
-  test('OHOS device smoke protects existing bundles', () {
-    const bundles = '''
-bundleName: com.example.other
-bundleName: com.example.smoke
-''';
-    expect(
-      ohosBundleAppearsInstalledForTesting(
-        bundles,
-        bundle: 'com.example.smoke',
-      ),
-      isTrue,
-    );
-    expect(
-      ohosBundleAppearsInstalledForTesting(
-        bundles,
-        bundle: 'com.example.smoke.extra',
-      ),
-      isFalse,
-    );
-  });
-
-  test('OHOS device smoke install never replaces an existing bundle', () {
-    expect(
-      ohosHdcInstallArgumentsForTesting(
-        deviceId: 'device-a',
-        hapPath: '/tmp/smoke.hap',
-      ),
-      ['-t', 'device-a', 'install', '/tmp/smoke.hap'],
-    );
-  });
-
-  test('OHOS device smoke validates the HAP bundle name', () {
-    expect(
-      ohosHapBundleNameForTesting('''
-{"summary":{"app":{"bundleName":"com.example.smoke"}}}
-'''),
-      'com.example.smoke',
-    );
-    expect(
-      () => ohosHapBundleNameForTesting('{"summary":{"app":{}}}'),
-      throwsFormatException,
-    );
-    expect(() => ohosHapBundleNameForTesting('[]'), throwsFormatException);
-  });
-
-  test('OHOS device smoke recognizes hdc outcomes and Rust marker', () {
-    expect(
-      ohosHdcInstallSucceededForTesting('install bundle successfully.'),
-      isTrue,
-    );
-    expect(
-      ohosHdcInstallSucceededForTesting('fail to verify pkcs7 file'),
-      isFalse,
-    );
-    expect(
-      ohosHdcAbilityStartSucceededForTesting('start ability successfully.'),
-      isTrue,
-    );
-    expect(
-      ohosDeviceSmokeLogPassedForTesting('''
-08-11 12:31:36.401 50093 50093 W Flutter: FRB_OHOS_SMOKE_RESULT=PASS
-''', expectedLog: 'FRB_OHOS_SMOKE_RESULT=PASS'),
-      isTrue,
-    );
-    expect(
-      ohosDeviceSmokeLogPassedForTesting(
-        '''
-08-11 12:30:00.500 40000 40000 W Flutter: FRB_OHOS_SMOKE_RESULT=PASS
-''',
-        baselineLogs: '''
-08-11 12:30:00.500 40000 40000 W Flutter: FRB_OHOS_SMOKE_RESULT=PASS
-''',
-        expectedLog: 'FRB_OHOS_SMOKE_RESULT=PASS',
-      ),
-      isFalse,
-    );
-    expect(
-      ohosDeviceSmokeLogPassedForTesting(
-        '''
-08-11 12:30:00.500 40000 40000 W Flutter: FRB_OHOS_SMOKE_RESULT=PASS
-08-11 12:31:36.401 50093 50093 W Flutter: FRB_OHOS_SMOKE_RESULT=PASS
-''',
-        baselineLogs: '''
-08-11 12:30:00.500 40000 40000 W Flutter: FRB_OHOS_SMOKE_RESULT=PASS
-''',
-        expectedLog: 'FRB_OHOS_SMOKE_RESULT=PASS',
-      ),
-      isTrue,
-    );
-    expect(
-      ohosDeviceSmokeLogPassedForTesting('''
-08-11 12:31:36.401 50093 50093 W Flutter: FRB_OHOS_SMOKE_RESULT=PASSIVE
-''', expectedLog: 'FRB_OHOS_SMOKE_RESULT=PASS'),
-      isFalse,
-    );
-    expect(
-      ohosDeviceSmokeLogPassedForTesting(
-        'FRB_OHOS_SMOKE_RESULT=PASS',
-        expectedLog: '   ',
-      ),
-      isFalse,
-    );
-  });
-
   test('GitHub release create command does not label prerelease versions', () {
     expect(
       githubReleaseCreateCommand(
